refactor: 공유큐 , 이미지 레포지토리 분리
